+++ app/gui2/package.json
@@ -66,7 +66,6 @@
     "pinia": "^2.1.7",
     "postcss-inline-svg": "^6.0.0",
     "postcss-nesting": "^12.0.1",
-    "reconnecting-ws": "^2.0.4",
     "rimraf": "^5.0.5",
     "semver": "^7.5.4",
     "sucrase": "^3.34.0",
