+++ app/gui2/shared/util/net/ReconnectingWSTransport.ts
@@ -1,3 +1,9 @@
+/**
+ * This file is modified version of open-rpc/client-js WebSocketTransport implementation
+ * (https://github.com/open-rpc/client-js/blob/master/src/transports/WebSocketTransport.ts)
+ * which uses the automatically reconnecting websocket.
+ */
+
 import { ERR_UNKNOWN, JSONRPCError } from '@open-rpc/client-js/build/Error'
 import {
   getBatchRequests,
@@ -6,19 +12,22 @@ import {
 } from '@open-rpc/client-js/build/Request'
 import { Transport } from '@open-rpc/client-js/build/transports/Transport'
 import WS from 'isomorphic-ws'
+import { ObservableV2 } from 'lib0/observable'
 import { WebSocket } from 'partysocket'
 
-class WebSocketTransport extends Transport {
+class ReconnectingWebSocketTransport extends Transport {
   public connection: WebSocket
   public uri: string
+  public events: ObservableV2<{ closed(event: CloseEvent): void }>
 
   constructor(uri: string) {
     super()
     this.uri = uri
-    this.connection = new WebSocket(uri, undefined, { WebSocket: WS })
+    this.connection = new WebSocket(uri, undefined, { WebSocket: WS, maxEnqueuedMessages: 0 })
+    this.events = new ObservableV2()
   }
   public connect(): Promise<any> {
-    return new Promise<void>((resolve, reject) => {
+    return new Promise<void>((resolve) => {
       const cb = () => {
         this.connection.removeEventListener('open', cb)
         resolve()
@@ -52,6 +61,23 @@ class WebSocketTransport extends Transport {
   public close(): void {
     this.connection.close()
   }
+
+  on<K extends keyof WebSocketEventMap>(
+    type: K,
+    cb: (
+      event: WebSocketEventMap[K] extends Event ? WebSocketEventMap[K] : never,
+    ) => WebSocketEventMap[K] extends Event ? void : never,
+  ): void {
+    this.connection.addEventListener(type, cb)
+  }
+  off<K extends keyof WebSocketEventMap>(
+    type: K,
+    cb: (
+      event: WebSocketEventMap[K] extends Event ? WebSocketEventMap[K] : never,
+    ) => WebSocketEventMap[K] extends Event ? void : never,
+  ): void {
+    this.connection.removeEventListener(type, cb)
+  }
 }
 
-export default WebSocketTransport
+export default ReconnectingWebSocketTransport
